- using temporary directory now for plot images, not upload directory - added check if plotted image actually exists - added marker functionality to line plots

+++ gui/panel/plot/image.php
@@ -9,12 +9,23 @@ class GUI_Panel_Plot_Image extends Scriptlet {
 	
 	public function display() {
 		header('Content-type: image/png');
-		$filename = ini_get('upload_tmp_dir').DS.$this->getParam('img');
-		$image = imagecreatefrompng($filename);
-		imagepng($image);
-		imagedestroy($image);
+		$filename = System::getTemporaryDirectory().DS.$this->getParam('img');
 		$file = new IO_File($filename);
-		$file->delete();
+		if ($file->exists()) {
+			$image = imagecreatefrompng($filename);
+			imagepng($image);
+			imagedestroy($image);
+			$file->delete();
+		}
+		else {
+			// TODO throw exception?
+			$image = imagecreate(110, 20);
+			$backgroundColor = imagecolorallocate($image, 0, 0, 0);
+			$textColor = imagecolorallocate($image, 255, 0, 0);
+			imagestring($image, 1, 5, 5,  'Couldn\'t load image', $textColor);
+			imagepng($image);
+			imagedestroy($image);
+		}
 	}
 }
